Refactor (Domain): Reorganize interfaces folder
