+++ src/api/MoRS.ManagementSystem.Domain/Interfaces/IEmailRepository.cs
@@ -1,0 +1,9 @@
+using MoRS.ManagementSystem.Domain.Entities;
+using MoRS.ManagementSystem.Domain.Interfaces.BaseInterfaces;
+
+namespace MoRS.ManagementSystem.Domain.Interfaces;
+
+public interface IEmailRepository : IBaseRepository<Email>
+{
+
+}
