+++ src/api/MoRS.ManagementSystem.Domain/Interfaces/IMalfunctionReportRepository.cs
@@ -1,6 +1,7 @@
 using MoRS.ManagementSystem.Domain.Entities;
+using MoRS.ManagementSystem.Domain.Interfaces.BaseInterfaces;
 
-namespace MoRS.ManagementSystem.Infrastructure.Interfaces.Repositories;
+namespace MoRS.ManagementSystem.Domain.Interfaces;
 
 public interface IMalfunctionReportRepository : IBaseRepository<MalfunctionReport>
 {
