+++ src/api/MoRS.ManagementSystem.Domain/Interfaces/IRoomRepository.cs
@@ -1,0 +1,9 @@
+using MoRS.ManagementSystem.Domain.Entities;
+using MoRS.ManagementSystem.Domain.Interfaces.BaseInterfaces;
+
+namespace MoRS.ManagementSystem.Domain.Interfaces;
+
+public interface IRoomRepository : IBaseRepository<Room>
+{
+
+}
