+++ src/api/MoRS.ManagementSystem.Domain/Interfaces/IUserRepository.cs
@@ -1,0 +1,9 @@
+using MoRS.ManagementSystem.Domain.Entities;
+using MoRS.ManagementSystem.Domain.Interfaces.BaseInterfaces;
+
+namespace MoRS.ManagementSystem.Domain.Interfaces;
+
+public interface IUserRepository : IBaseRepository<User>
+{
+
+}
