--- src/api/MoRS.ManagementSystem.Domain/Interfaces/Repositories/IEmailRepository.cs
@@ -1,8 +1,0 @@
-using MoRS.ManagementSystem.Domain.Entities;
-
-namespace MoRS.ManagementSystem.Infrastructure.Interfaces.Repositories;
-
-public interface IEmailRepository : IBaseRepository<Email>
-{
-
-}
