--- src/api/MoRS.ManagementSystem.Domain/Interfaces/Repositories/IRoleRepository.cs
@@ -1,8 +1,0 @@
-using MoRS.ManagementSystem.Domain.Entities;
-
-namespace MoRS.ManagementSystem.Infrastructure.Interfaces.Repositories;
-
-public interface IRoleRepository : IBaseRepository<Role>
-{
-
-}
