--- src/api/MoRS.ManagementSystem.Domain/Interfaces/Repositories/IUserRepository.cs
@@ -1,8 +1,0 @@
-using MoRS.ManagementSystem.Domain.Entities;
-
-namespace MoRS.ManagementSystem.Infrastructure.Interfaces.Repositories;
-
-public interface IUserRepository : IBaseRepository<User>
-{
-
-}
