+++ src/api/MoRS.ManagementSystem.Infrastructure/Extensions/RepositoryServiceCollectionExtensions.cs
@@ -1,5 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
-using MoRS.ManagementSystem.Infrastructure.Interfaces.Repositories;
+using MoRS.ManagementSystem.Domain.Interfaces;
 using MoRS.ManagementSystem.Infrastructure.Repositories;
 
 namespace MoRS.ManagementSystem.Infrastructure.Extensions;
